+++ test/pleroma/web/mastodon_api/views/account_view_test.exs
@@ -90,7 +90,11 @@ defmodule Pleroma.Web.MastodonAPI.AccountViewTest do
         hide_follows_count: false,
         relationship: %{},
         skip_thread_containment: false,
-        accepts_chat_messages: nil
+        accepts_chat_messages: nil,
+        email_notifications: %{
+          "digest" => false,
+          "notifications" => ["mention", "pleroma:chat_mention"]
+        }
       }
     }
 
@@ -190,7 +194,11 @@ defmodule Pleroma.Web.MastodonAPI.AccountViewTest do
         hide_follows_count: false,
         relationship: %{},
         skip_thread_containment: false,
-        accepts_chat_messages: nil
+        accepts_chat_messages: nil,
+        email_notifications: %{
+          "digest" => false,
+          "notifications" => ["mention", "pleroma:chat_mention"]
+        }
       }
     }